refactor: 딜러와 플레이어 추상화 적용
