+++ src/test/java/domain/DealerTest.java
@@ -17,19 +17,6 @@ class DealerTest {
     }
 
     @Test
-    void 카드의_합을_계산한다() {
-        CardDeck cardDeck = new CardDeck();
-        Dealer dealer = new Dealer(cardDeck);
-
-        dealer.add(카드(Denomination.KING));
-        dealer.add(카드(Denomination.SIX));
-
-        int result = dealer.calculateScore();
-
-        assertThat(result).isEqualTo(16);
-    }
-
-    @Test
     void 점수의_합이_17보다_작으면_카드를_더_받아야_한다() {
         CardDeck cardDeck = new CardDeck();
         Dealer dealer = new Dealer(cardDeck);
